Extract enum
